chore: rename `bench` file
